Finished first iteration for Custom object support
