--- richtextfx-demos/sample2.rtfx
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<StyledDocument>
-  <Paragraph style="">
-    <StyledText style="-fx-font-size: 12pt;-fx-font-family: Serif;-fx-fill: rgb(0, 0, 0);">Lorem Ipsum blafasel</StyledText>
-  </Paragraph>
-  <Paragraph style="-fx-text-alignment: center;">
-    <StyledText style="-fx-font-size: 12pt;-fx-font-family: Serif;-fx-fill: rgb(0, 0, 0);">sdfsf sdfd sdfdsf </StyledText>
-  </Paragraph>
-  <Paragraph style="">
-    <StyledText style="-fx-font-size: 12pt;-fx-font-family: Serif;-fx-fill: rgb(0, 0, 0);">sdf ewrewr we</StyledText>
-    <InlineImage fileName="file:/C:/Users/afester/Projects/RichTextFX/richtextfx-demos/src/main/resources/org/fxmisc/richtext/demo/richtext/underline.png"/>
-    <StyledText style="-fx-font-size: 12pt;-fx-font-family: Serif;-fx-fill: rgb(0, 0, 0);">rwfcxvcxv</StyledText>
-  </Paragraph>
-  <Paragraph style="">
-    <StyledText style="-fx-font-size: 12pt;-fx-font-family: Serif;-fx-fill: rgb(0, 0, 0);">tztz gf vcb </StyledText>
-    <StyledText style="-fx-font-size: 12pt;-fx-font-family: Serif;-fx-fill: rgb(0, 0, 0);-rtfx-background-color: rgb(204, 51, 51);">zth</StyledText>
-    <StyledText style="-fx-font-size: 12pt;-fx-font-family: Serif;-fx-fill: rgb(0, 0, 0);"> gfjfgf</StyledText>
-  </Paragraph>
-  <Paragraph style="">
-    <StyledText style="-fx-font-size: 12pt;-fx-font-family: Serif;-fx-fill: rgb(42, 55, 1);">ffsdfdds </StyledText>
-    <InlineImage fileName="file:/C:/Users/afester/Projects/RichTextFX/richtextfx-demos/src/main/resources/org/fxmisc/richtext/demo/richtext/savefile.png"/>
-    <InlineImage fileName="file:/C:/Users/afester/Projects/RichTextFX/richtextfx-demos/src/main/resources/org/fxmisc/richtext/demo/richtext/cut.png"/>
-    <StyledText style="-fx-font-size: 12pt;-fx-font-family: Serif;-fx-fill: rgb(42, 55, 1);">sfs</StyledText>
-  </Paragraph>
-  <Paragraph style="">
-    <StyledText style=""/>
-  </Paragraph>
-  <Paragraph style="">
-    <StyledText style=""/>
-  </Paragraph>
-  <Paragraph style="">
-    <StyledText style=""/>
-  </Paragraph>
-  <Paragraph style="">
-    <StyledText style=""/>
-  </Paragraph>
-  <Paragraph style="">
-    <StyledText style=""/>
-  </Paragraph>
-  <Paragraph style="">
-    <StyledText style=""/>
-  </Paragraph>
-</StyledDocument>
