--- richtextfx-demos/src/main/resources/org/fxmisc/richtext/demo/textstructure-example.css
@@ -1,46 +1,0 @@
-
-.richtext {
-  -fx-background-image: url(background2.png);
-  -fx-font-family: serif;
-}
-
-
-/* normal paragraph */
-.p {
-  -fx-font: 12pt "Sans";
-  /* -fx-padding: 0px 0px 12px 10px; */
-}
-
-/* Header 1 paragraph */
-.h1 {
-  -fx-font: 20pt "Sans";
-  /*-fx-fill: green;*/
-  /*-fx-padding: 0px 0px 12px 0px;*/
-}
-
-.h2 {
-  -fx-font: 16pt "Sans"; 
-}
-
-.java {
-  -fx-font: 10pt "Courier New";
-  -fx-background-color: #f0f0f0;
-}
-
-
-/* Inline styles */
-
-.em {
-  -fx-font-weight: bold;
-}
-
-.strong {
-  -fx-fill: yellow;
-  -fx-backgrund-color: blue;
-  -fx-stroke: green;
-}
-
-
-.keyword {
-  -fx-fill: red;
-}
