+++ richtextfx/src/test/java/org/fxmisc/richtext/model/CustomObjectTest.java
@@ -2,21 +2,23 @@ package org.fxmisc.richtext.model;
 
 import static org.junit.Assert.*;
 import org.junit.Test;
+import static org.hamcrest.CoreMatchers.instanceOf;
+import static org.hamcrest.MatcherAssert.assertThat;
 
 public class CustomObjectTest {
 
     @Test
-    public void testCustomObjectCreation() {
+    public void testLinkedImageCreation() {
         SimpleEditableStyledDocument<Boolean, String> doc = 
                 new SimpleEditableStyledDocument<>(true, "");
         ReadOnlyStyledDocument<Boolean, String> customObj = 
-                ReadOnlyStyledDocument.createObject(new InlineImage("", "sample.png"), true, "");
+                ReadOnlyStyledDocument.createObject(new LinkedImage<String>("", "sample.png"), true, "");
         doc.replace(0, 0, customObj);
         assertEquals(1, doc.getLength());
 
         Paragraph<Boolean, String> para = doc.getParagraphs().get(0);
         Object x = para.getSegments().get(0);
-        assertTrue(x instanceof CustomObject);
+        assertThat(x, instanceOf(LinkedImage.class));
     }
 
 }
